Switch construct to satoshis

+++ service/construct/args.py
@@ -4,7 +4,7 @@ from .. import constants
 
 class BuildArgs(BaseModel):
     receive_address: Union[str, None] = Field(default=None)
-    marker: float = Field(default=constants.DEFAULT_MARKER)
-    fee: float = Field(default=constants.DEFAULT_FEE)
+    marker: int = Field(default=constants.DEFAULT_MARKER)
+    fee: int = Field(default=constants.DEFAULT_FEE)
     send_address: str
     payload: str
